requirements.txt
pytrends==4.9.2
google-generativeai==0.5.2
requests==2.31.0
markdown==3.6
python-dotenv==1.0.1
tweepy==4.14.0
